requirements.txt
streamlit==1.32.2
google-generativeai==0.4.1